+++ apps/creator-app/src/components/Tip/TipHotkey/TipHotkey.css.ts
@@ -1,0 +1,10 @@
+import { style } from '@voiceflow/style';
+
+export const iconStyle = style({
+  width: 28,
+  height: 28,
+  marginLeft: -9,
+  marginTop: -9,
+  marginBottom: -9.5,
+  marginRight: -8,
+});