+++ apps/creator-app/src/components/Tip/TipPortal/TipPortal.component.tsx
@@ -6,7 +6,12 @@ import { Transition } from 'react-transition-group';
 import { useLocalStorageState } from '@/hooks/storage.hook';
 
 import { containerStyle } from './TipPortal.css';
-import { ITipPortal } from './TipPortal.interface';
+
+interface ITipPortal {
+  scope: string;
+  closing?: boolean;
+  children: (props: { onClose: VoidFunction }) => React.ReactNode;
+}
 
 export const TipPortal: React.FC<ITipPortal> = ({ scope, closing, children }) => {
   const [visible, setVisible] = useLocalStorageState(`vf_tip:${scope}`, true);