+++ apps/creator-app/src/pages/AssistantCMS/pages/CMSEntity/components/CMSEntityTable/CMSEntityTable.component.tsx
@@ -6,7 +6,11 @@ import { CMS_ENTITY_LEARN_MORE } from '@/constants/link.constant';
 import { EMPTY_TEST_ID, TABLE_TEST_ID } from '@/pages/AssistantCMS/AssistantCMS.constant';
 
 import { CMSEmpty } from '../../../../components/CMSEmpty/CMSEmpty.component';
-import { useCMSRowItemClick, useCMSRowItemContextMenu, useCMSRowItemNavigate } from '../../../../hooks/cms-row-item.hook';
+import {
+  useCMSRowItemClick,
+  useCMSRowItemContextMenu,
+  useCMSRowItemNavigate,
+} from '../../../../hooks/cms-row-item.hook';
 import { useEntityCMSManager, useOnEntityCreate } from '../../CMSEntity.hook';
 import { entityColumnsOrderAtom } from './CMSEntityTable.atom';
 import { CMS_ENTITY_TABLE_CONFIG } from './CMSEntityTable.config';
@@ -16,13 +20,19 @@ export const CMSEntityTable: React.FC = () => {
   const onCreate = useOnEntityCreate();
   const onRowClick = useCMSRowItemClick();
   const onRowNavigate = useCMSRowItemNavigate();
-  const rowContextMenu = useCMSRowItemContextMenu({ nameColumnType: EntityTableColumn.NAME });
+  const rowContextMenu = useCMSRowItemContextMenu({
+    nameColumnType: EntityTableColumn.NAME,
+  });
   const entityCMSManager = useEntityCMSManager();
 
   return (
     <CMSEmpty
       title="No entities exist"
-      button={{ label: 'Create entity', onClick: (search) => onCreate({ name: search }), testID: tid(EMPTY_TEST_ID, 'create-entity') }}
+      button={{
+        label: 'Create entity',
+        onClick: (search) => onCreate({ name: search }),
+        testID: tid(EMPTY_TEST_ID, 'create-entity'),
+      }}
       searchTitle="No entities found"
       description="Entities help your assistant know which data to pluck out from the users response. "
       illustration="NoContent"