+++ libs/orm-designer/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.62.2](https://github.com/voiceflow/creator-app/compare/@voiceflow/orm-designer@2.62.1...@voiceflow/orm-designer@2.62.2) (2024-05-24)
+
+**Note:** Version bump only for package @voiceflow/orm-designer
+
+## [2.62.1](https://github.com/voiceflow/creator-app/compare/@voiceflow/orm-designer@2.62.0...@voiceflow/orm-designer@2.62.1) (2024-05-24)
+
+**Note:** Version bump only for package @voiceflow/orm-designer
+
+# [2.62.0](https://github.com/voiceflow/creator-app/compare/@voiceflow/orm-designer@2.61.5...@voiceflow/orm-designer@2.62.0) (2024-05-23)
+
+### Features
+
+* add gpt-4o to AI model dropdown and settings (ML-892) ([#8147](https://github.com/voiceflow/creator-app/issues/8147)) ([cb18db7](https://github.com/voiceflow/creator-app/commit/cb18db7e3593c9032dedd33553ba6e53d2f5d4f1))
+
 ## [2.61.5](https://github.com/voiceflow/creator-app/compare/@voiceflow/orm-designer@2.61.4...@voiceflow/orm-designer@2.61.5) (2024-05-22)
 
 **Note:** Version bump only for package @voiceflow/orm-designer