+++ libs/realtime-sdk/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.402.1](https://github.com/voiceflow/creator-app/compare/@voiceflow/realtime-sdk@1.402.0...@voiceflow/realtime-sdk@1.402.1) (2024-05-24)
+
+### Bug Fixes
+
+* revert integrate org members (COR-2143) ([#8172](https://github.com/voiceflow/creator-app/issues/8172)) ([6f8c05c](https://github.com/voiceflow/creator-app/commit/6f8c05c04d076e6e995f0b83ecdd9861869b1729))
+
+# [1.402.0](https://github.com/voiceflow/creator-app/compare/@voiceflow/realtime-sdk@1.401.1...@voiceflow/realtime-sdk@1.402.0) (2024-05-24)
+
+### Features
+
+* integrate org members (COR-2143) ([#8160](https://github.com/voiceflow/creator-app/issues/8160)) ([ea27b1b](https://github.com/voiceflow/creator-app/commit/ea27b1b3516953427a0433fcb50380d77b362f5b))
+
+## [1.401.1](https://github.com/voiceflow/creator-app/compare/@voiceflow/realtime-sdk@1.401.0...@voiceflow/realtime-sdk@1.401.1) (2024-05-23)
+
+### Bug Fixes
+
+* invalid node reference during migration (COR-2147) ([#8168](https://github.com/voiceflow/creator-app/issues/8168)) ([b3b597e](https://github.com/voiceflow/creator-app/commit/b3b597e4f7cf4db1b5ef709ed681088a2628c2cd))
+
+# [1.401.0](https://github.com/voiceflow/creator-app/compare/@voiceflow/realtime-sdk@1.400.1...@voiceflow/realtime-sdk@1.401.0) (2024-05-23)
+
+### Features
+
+* add gpt-4o to AI model dropdown and settings (ML-892) ([#8147](https://github.com/voiceflow/creator-app/issues/8147)) ([cb18db7](https://github.com/voiceflow/creator-app/commit/cb18db7e3593c9032dedd33553ba6e53d2f5d4f1))
+
 ## [1.400.1](https://github.com/voiceflow/creator-app/compare/@voiceflow/realtime-sdk@1.400.0...@voiceflow/realtime-sdk@1.400.1) (2024-05-22)
 
 ### Bug Fixes